+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo.hs
@@ -104,6 +104,7 @@ mkAlonzoStAnnTx ei sysStart pp utxo tx =
         , ltiSystemStart = sysStart
         , ltiUTxO = utxo
         , ltiTx = tx
+        , ltiMemoizedSubTransactions = mempty
         }
    in
     AlonzoStAnnTx